Add support for go.
Each font is represented by a Font-typed variable in the IconFontCppHeaders
namespace, where Font is a small struct that wraps up the filenames and
range of unicode values.

Unlike the C++/C# bindings, the name->unicode mappings for each glyph are
stored in a map, rather than in a flat set of defined constants. There is
thus some additional overhead in looking up a given mapping, though with
some flexibility on the user side in return (ability to iterate over the
mappings, etc.)